refactor(kb-load): extract common sanity tests to sibling file
primer-kb-load/tests/common/mod.rs carried a single inline
`#[cfg(test)] mod sanity_tests` (10 benchmark-shape sanity tests over the
EN/DE retrieval query sets). Move it verbatim into a sibling
`tests/common/sanity_tests.rs` and replace the inline block with
`#[cfg(test)] mod sanity_tests;`.

common/mod.rs is a `mod.rs`-style module, so the submodule resolves to a
sibling file in the same dir (no mkdir). The test module is a descendant,
so `use super::*` reaches all parent items with no visibility edits. No
depth-sensitive paths in the body (no include_str!/crate::), so the cut is
fully verbatim. First non-`src/` cut in the test-extraction series.

mod.rs 791 -> 677 lines (-114); new sanity_tests.rs 113 lines.

Verification:
- cargo test -p primer-kb-load sanity_tests -> 10 passed in each of 7
  active binaries, identical to baseline.
- cargo test --workspace -> 0 failures (51 ok result blocks).
- cargo clippy -p primer-kb-load --all-targets -D warnings -> clean.
- cargo fmt --all -- --check -> clean.

Test-only relocation; no behaviour change, no incidental fixes.
README/ROADMAP carry no reference to this file.
